PIM-9356: Fix external api endpoint for products models with invalid quantified associations
