+++ tests/back/Pim/Enrichment/Integration/Storage/Sql/ProductModel/QuantifiedAssociation/GetProductQuantifiedAssociationsByProductModelCodesIntegration.php
@@ -9,6 +9,7 @@ use Akeneo\Pim\Enrichment\Component\Product\Model\QuantifiedAssociation\Quantifi
 use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
 use AkeneoTest\Pim\Enrichment\EndToEnd\Product\EntityWithQuantifiedAssociations\QuantifiedAssociationsTestCaseTrait;
 use AkeneoTest\Pim\Enrichment\Integration\Storage\Sql\AbstractQuantifiedAssociationIntegration;
+use Doctrine\DBAL\Connection;
 
 class GetProductQuantifiedAssociationsByProductModelCodesIntegration extends AbstractQuantifiedAssociationIntegration
 {
@@ -297,6 +298,32 @@ class GetProductQuantifiedAssociationsByProductModelCodesIntegration extends Abs
         $this->getAssociationTypeRemover()->remove($associationType);
         $actual = $this->getQuery()->fromProductModelCodes(['productModelA']);
 
+        $this->assertSame([], $actual);
+    }
+
+    /**
+     * @test
+     *
+     * https://akeneo.atlassian.net/browse/PIM-9356
+     */
+    public function itDoesNotFailOnInvalidQuantifiedAssociations()
+    {
+        $rootProductModel = $this->getEntityBuilder()->createProductModel('root_product_model', 'familyVariantWithTwoLevels', null, []);
+        $subProductModel = $this->getEntityBuilder()->createProductModel('sub_product_model_1', 'familyVariantWithTwoLevels', $rootProductModel, []);
+        $this->getEntityBuilder()->createVariantProduct('variant_product_1', 'aFamily', 'familyVariantWithTwoLevels', $subProductModel, []);
+
+        /** @var Connection $connection */
+        $connection = $this->get('doctrine.dbal.default_connection');
+
+        $query = <<<SQL
+        UPDATE pim_catalog_product_model
+        SET quantified_associations = '[]'
+        WHERE code = 'root_product_model'
+SQL;
+
+        $connection->executeUpdate($query);
+
+        $actual = $this->getQuery()->fromProductModelCodes(['variant_product_1']);
         $this->assertSame([], $actual);
     }
